Fix Lazy.groupBy again
The problem was happening for operators that are not equivalence relations
eg (<). One bug was that it was using the operator the wrong way wround.
The more subtle problem was that it was using the final element in the
group to compare against rather than the first as Data.List.groupBy does.
For example groupBy (<) [0,10,5]

darcs-hash:20060517170250-9561d-623227f48a87dc1186d20493c2fdbc57aca5f794.gz