Change the implementation of the unfoldr(N) functions.
Use a more compact implementation for unfoldrN and change it's behaviour
to only return Just in the case that it actually 'overflowed' the N, so
the boundary case of unfolding exactly N gives Nothing. This also helps
with maintaining the invariant in .Lazy
Implement unfoldrr and Lazy.unfoldr in terms of unfoldrN. Use fibonacci
growth for the chunk size in unfoldr and simple quadratic growth for
Lazy.unfoldr and in the latter case an upper bound on the chunk size.
upper

darcs-hash:20060524044732-9561d-041bb56c9e5c896a5b44dd2d21d6d3fcf05eb3f1.gz