Swap the result tuple in loopU
because it's obviously the wrong way round! ;-)

Currently we have:
loopU :: (acc -> Word8 -> (acc, Maybe Word8))
      -> acc
      -> ByteString
      -> (ByteString, acc)
which has the result of the per-elem map/fold/filter func returning
(acc, Maybe Word8) but the overall result as (ByteString, acc) which
is reversed. So we also have to swap loopArr, loopAcc and loopSndAcc.
At the same time I'm generalising their types so that we'll be able
to use them (and the rules that apply to them) to [ByteString] as
well as just ByteString.

darcs-hash:20060524160046-9561d-2668e6d6118636cb095acfc57429740445379a78.gz