Add lazy-as-available byte strings
Lazy reading of network sockets is fantastic, but often you need to process
data as it's sent. Currently you can either set the block size to 1 byte or
you have to wait for an entire block to be read.

This adds the (terribly named) hGetSomeContentsN, which returns the contents
of a handle, lazily, but in a timely fashion.

darcs-hash:20060813194125-19166-914c5c380f26c556caabff5be7dc39a08f35ca4e.gz