Implement byteswapping, used for endian fiddling.
If the ByteString is going to end up being interpreted as a binary
representation of a sequence of 4-byte quantities (e.g. Float), then one
might need to change the endianness (byte ordering) of the
representation.  This patch adds the API call
   byteswap :: ByteString -> ByteString
for such applications.  It assumes the bytestring will be a multiple of
4 in size.  (If not, the last n<4 bytes will be untouched.)  It also
assumes (without checking) that the start of the bytestring has the
correct alignment.

darcs-hash:20070227163230-6871e-e8c0dd7301127647be6a7bb0b3c761fcecac259c.gz