hGetContents: use hGet instead of hGetNonBlocking + hWaitForInput + hIsEOF
Ignore-this: 43bac03fc6d7de74564f5cd608ed9f17

Not only is this cleaner, but it fixes a problem with read binary data
over a pipe, see
  http://hackage.haskell.org/trac/ghc/ticket/3808

The problem is that bytestring normally works fine when the Handle is
not in binary mode, because it uses hGetBuf/hPutBuf which bypass the
encoding layer.  That is, except in this one particular case:
hWaitForInput might do some decoding, because it has to determine
whether there are any characters (not bytes) available to be read.

I imagine hGetNonBlocking was used due to concerns that hGet might
block if there is insufficient data, but that's not the case; hGet
returns a short read if it would otherwise block, and only blocks if
there is no data to read.

darcs-hash:20100322153914-12142-6652e014868865f74a83c4304989c8b5d13a7505.gz