introduce 'Builder.Extras' module
this module holds all the functions that more experienced users can
use to fine-tune the performance of their 'Builder's. Specifically,

  - bytestring insertion and copying
  - execution of Builders with different allocation strategies
  - host-specific encoding of integers

The later is also along the design principle that non-experienced
users should be spared of thinking about portability: we just don't
expose functions that could lead to non-portability in their interface.