Adapt to renaming of 'System.IO.Write' to 'Codec.Bounded.Encoding'
The documentation is only partially updated, but it is becoming much
cleaner. Functions like 'encodeWith' are now quite self-explanatory.
The plan is to have a single module 'Codec.Bounded.Encoding' in the
bytestring library that exports all the functions for defining and
executing bounded encodings.

An additional advantage of making bounded encodings first-class citizens
of the bytestring library is that most functions of type 'a -> Builder'
become convenience functions. Hence, we only have to keep the most
important onces.  The others can be constructed by the library user
himself.

As a first step towards this direction, I dropped the ASCII encoding
builders. Library users that really need them can construct them by
themself and others are insured from sudden errors due to unencodable
characters.