benchmarked decimal encoding of 'Double's
This includes comparing to a ported version of 'blaze-textual'. The conclusion
is that the bottleneck lies in the conversion from 'Double' to its decimal
representation and not in the conversion of this decimal representation to the
sequence of bytes. We thereore use Haskell's built-in conversion, as we were
not able to improve upon it. For high-performance use, the double-conversion
package can be incorporated easily by wrapping it as a 'BoundedEncoding'.